EMCal CorrFW: Refactor branch creation
Refactor track and cluster branch creation and configuration,
as well as simplify the cell creation and copying process.

Note: We only create inputs for requested objects. This is not
particularly important for cluster and track containers since they
cause little overhead, but it is extremely important for cells
since there is substantial overhead