Workaround C++11 in FastJet 3.2 with ROOT<6
ROOT < 6 does not support C++11. While it's fine to have C++11 code in ROOT
class implementations it is not supported in headers due to dictionary
generation.  Newer FastJet versions (notably v3.2.1) make use of the C++11
`override` keyword which we make mute only during dictionary generation in ROOT
versions < 6. Note that this is a workaround.