Macros to create TPCRecoParams as used in 2015 data with SCD
MakeTPCRecoParam_Run2Dist.C creates the recoparam for Low/High flux data with distortions, as used
for reconstruction of LHC15o, LHC15n and LH16x... For the cosmics the usage of CorrectionMaps is not
used, instead Composed correction will be used. Note that some cosmic runs of 2016 used CorrectionMaps,
while 2015 cosmics reco was done with ComposedCorrections.
MakeTPCRecoParam_Run2DistMC.C provides the recoparams to be used for MC with distortions: exactly
the same as for raw data except that the usage of IonTail and CrossTalk corrections is OFF.