Add task checking the influence of noisy fastors to the patch
In addition to the total patch ADC the single fastors contribution to
a trigger patch are investigated. Probes are typically the highest energy
fastor, the number of non-zero fastors as well as the sum of good and masked
fastor ADCs. Maximum counts are shown relative to the total patch ADC.