Task inspecting triggers rejected as noise triggers
This task specially analyses events rejected as noise triggers
and fills histograms sensitve at FastOR and at patch level. In
addition the class AliEMCALTriggerPatchADCInfo is now shared
among different tasks and moved to PWG/EMCAL/EMCALTrigger. In
preparation of the move to AliRoot it got the ending tag AP
in order to prevent multiple symbols.