Partial implementation of the MFT Support and PCBs
The MFT support implemented in this commit is not definitive,
as the final MFT support geometry is not defined yet. The Support
corresponds to STP files from November 2015 and have not been
completely verified nor checked for overlaps.

Details at https://git.cta.if.ufrgs.br/pezzi/ALICE_MFT_UFRGS