EMCal CorrFW: Loop over tracks, cluster containers
All correction components (except for the clusterizer) can loop over
multiple cluster and track containers. This approach is more flexible
and is needed for embedding (although it is not specific to it).

This commit also includes:
- Updated to use iterators while looping over container contents. This
  should be a bit safer.
- Some fixes for index mapping

NOTE: The Hadronic Correction needed a bit of a hack. In some cases
including multiple containers in AODs, only the particle acceptance
conditions from the first particle container are applied. A warning is
thrown when these conditions occur. See the code for more details.