EMCal CorrFW: Fix run period option, other fixes
Includes:
- Option to set the track cuts period for each individual track
  container
- Bug fix in setting a blank run period

Note on selecting the run period:

It is strongly recommended that the user set the run period in the
correction task and in their user configuration! This is to ensure that
the user does not accidentally use a configuration from the wrong period
without noticing!

Although not at all recommended, the run period check can be disabled by
passing an empty string to SetRunPeriod() and in the user configuration.
A bug prevented this from working previously.

Also fixes for:
- pass detection
- YAML parsing node name check