Update LumiTools LUT for pPb
The LUT for ref.trigger class name, xsection is completed by pPb values. The COG values stored in the maps
of 5 TeV pPb are still using pp x-sections. For this reason they will be ignored (via default value of
new flag fUseMapLumiInfoCOG=kFALSE in the TPC recoparam) and will be recalculated on the fly