handling of event weights in both task and base  class
after new event selection criteria for 15o data, min bias centrality distribution is no longer flat in central collisions. this patch provides supplying weghts in the form of a TH1 containing a centrality distribution with arbitrary binning. weights are propagated both to the phi v2 class and to the base classes, a histogram with applied weight s- derived from the input distribution, is stored in the output list of the phi distribution. a sane default is provided in the addtask. to launch the weighting procedure, all the user has to do is add the histogram wigh the cen distribution via 'SetCentralityWeights()'

fix two is propagating the tpc q-vectors to the base class so that all necessary correlators are avaiable in the base class output