EMCal CoreFW: Add EMCal container utils
The container utils static class consists of helper functions for
AliEmcalContainer derived classes and related tasks.

This change also updates classes that rely on the functions now included
in the container utils static class.

Implementing as a namespaace causes problems with ROOT. As I understand
it, rootmaps cannot load namespaces automatically, so every function
would have to be added to the LinkDef explicitly, which isn't great.
Instead, static methods in class serve basically the same purpose