Add option to trigger on clusters instead of patches
New option applying trigger selection on clusters instead of patches.
Energy / ET is taken from clusters instead of patches once the option
is activated. Energy source is steered via the energy definition:
If energy definition is kClusterEnergy or kClusterTransverseEnergy then
the energy is taken from clusters.

The interface is changed in a way that now AliVEvents are used as function
argument in IsOfflineSelected. A cluster container name is mandatory
for the triggr on cluster energies.