Define smart pointer handling cut results
In case further information needs to be passed
from the track selection, the return value
bool is not sufficient. On the other side, the
original object must be unmodified.

Smart pointers are a reasonable tool to pass
further information from the track selection to
user codes. This smart pointer contains besides
the physical pointer to the original object also
information about the selection status and
a long value which can be handled by the user.
Thanks to overloaded operator bool, returning
the selection status, the smart pointer can
be used like a normal boolean return value
of a cut class. Access to members of the
original track is made easy by overloading
operators * and ->.

Note: Smart pointer in this case does not manage
memory.