EMCal CorrFW: Improve enums
Improvements include:
- Cleanup of enum maps to follow better coding conventions
- Simplify InputCellType:

This update simplifies InputCellType by removing or simplifying many of
the options. It was then renamed to reflect that it only affected
embedding options. InputCellType was composed of 4 classes of input data
for clustering:
- Non-Embedded (standard data)
- Embedded -> Selection of whether to include or exclude the energy from
  the MC embedded cells
- Pattern -> Used for generating a test pattern to test the clusterizer
- FastOR -> Superseded by the TriggerMaker

Non-Embedded is now the default mode, the embedded options are still
available, the pattern test mode is now a switch in the YAML
configuration, and the FastOR options are removed.