EMCal CorrFW: Update configuration printing
Previously relied on printing at the AliError level to ensure that the
configuration would be printed even on the LEGO train. However, this
will show "error" right next to the message, potentially causing
confusion since it wasn't an error. Switched to cout since it is not
filtered on the train and won't have the confusing message.