EMCal CorrFW: Handle default pass for MC
The warning repeats for each component, which is a bit annoying. This
will be resolved more carefully with a future commit.

This commit also creates a set of sample configurations for processing
MC. The main differences is to set the pass to "default" and to disable
the time and energy calibration, as we were told should be disabled.