Add support for user-defined pt-hard binning
The amount of pt-hard bins was hard coded to 11. With
this commit users can set a different amount of pt-hard
bins, still keeping the default value at 11 in order
to be backward compatible with most of the analyses.
When running with a non-default binning, in addition
to the amount of bins also bin limits should be defined
by the user in a linear list with size nbins + 1,
otherwise bin labels for the histograms in pt-hard
bins will not be shown.

With this version the MC dataset LHC16c2 (PYTHIA8
pt-hard production for pp 8 TeV 2012) with a non-
defalt pt-hard binning will be supported by the
AliAnalysisTaskEmcal.