Fix for potential (and actual) invalid memory access
The mass fitter should NOT be deleted if the fit results are drawn in a canvas.
Prior to this fix the canvas with the drawn fit results would crash the program as soon as a redraw is triggered because it cannot find anymore the drawn objects.
In this case the mass fitter is stored in a vector and deleted only in the destructor of the class.