EMCal Embed: Create init for task
The embedding task must now be initialized in the run macro or wagon.
This change centralizes the determination of the filenames to embed,
decreasing the load on AliEn, as well ensuring that the filenames are
the same for every job (that already should have been the case, but this
makes it certain).