Fix buggy initialization of pt-hard information
In local tests it turned out that the event is not
fully initialized at the step of UserNotify, in
particular the run number is not set.

In order to make the per-file initialization robust
the step was moved to a new function FileChanged (with
UserFileChanged for User-per-file initializations).
Unlike UserNotfiy this function is called at the
beginning of UserExec only in case the fFileChanged
signal (implemented as member variable) is set by
the UserNotify function. With this order the event
is guaranteed to be fully initialized.