EMCal Embed: Fix vector out-of-range
When looping around at the end of a file list, the pythia cross section
filenames vector went out-of-range when before the TChain catches that
it has finished the last file. The range is now explicitly checked and
an error is thrown that should match up with the end of file list error.