Enable optional CVMFS OCDB usage
This feature is required for accessing the OCDB from CVMFS expecially during the
MC release validation. OCDB is automatically accessed from there, if:

* `AliCDBManager::SetDefaultStorage("raw://")`
* and environment variable `OCDB_PATH` is set to a proper CVMFS path, e.g.
  /cvmfs/alice-ocdb.cern.ch

Run mapping to the correct folder is performed the same way as on AliEn, by
using the same XML mapping file. Mapping is not performed in case
`SetDefaultStorage` is set to a specific directory.

Those changes are backwards-compatible: the previous behaviour is unchanged in
case the `OCDB_PATH` variable is not exported. CVMFS OCDB must be enabled
explicitly.