EMCal Embed: Remove obsolete code
Removed code used to select only MC or data energy in the old embedding
framework. This code isn't used, but the options can cause user
confusion.

In the new embedding framework, either type can be selected by simpling
running the clusterizer with only the MC or data cells given
as the input to the clusterizer (instead of the combined cell collection).