Fixed V0M as a default for Run2 pPb
So far AliMultSelection task doesn't provide "V0A" we can use only "V0M". This is inconsistent for Run1 pPb.  However, AliCentrality task doesn't work for Run2 pPb, so it's not ideal but we decide to use V0M as a default. This would be fixed when V0A is available.