BC, Warning due to ambiguity in run range fixed
OADB was fixed due to an ambiguity warning for run range 246424-246846. The result will be the same, since the OADB code always get the latest object, but a warning message happened when someone loaded the BC for a run in the range 246424-246846, due to a previous object set there. The new object has a wider range which covers this range as well.