Make option to switch on long-lived decayse in PYTHIA8
Decays of long-lived particles (K0short, lambda) are currently
disabled in a hard coded way in the AliPythia8 wrapper. While
this is OK for normal MC productions, on the MC_gen lego train
this might not always be the desired behaviour. The patch allows
for an optional switch on of the decayer for long lived particles
(default: off)