Fix merge from .zip not from AliEn
Merging .zip files from non-AliEn paths requires specifying the inner .root file
with an anchor, in the format `arch.zip#file.root`. On AliEn this is not
necessary as the .zip entries are stored in the catalog as part of the current
directory.

It also tells the TFileMerger to access the .zip files not on AliEn directly, as
opposed to copying them locally first, as `TFile::Cp()` does not support .zip
anchors (silently).