Removed old EvaluateTriggerLogic
The new Logic does not rely on the _very_ slow TFormula. Instead it
does the job by creating a vector of booleans; one bool for each token
in the trigger logic (e.g. "V0A"). At the end, the final trigger
decission is true if all elements in that vector were true (taking
possible negations ("!") into account. If one or more are false, this
logic yields false.

I tried this new implemenation on LHC15o and LHC16k AODs without
issues and the results were identical to the old implementation. The
Efficiency gain of this new implemntation is several orders of
magnitude. This code is called many times per event and seems to
be responible for most of the CPU consumption. See issue #1507