EMCal Embed: Fix embedded event selection
Event selection was being applied to the input event, not the embedded
event.

In updating the event, it became clear that ESD event selection is not
possible at the moment due to not creating an input event handler. In
the interest of simplicity and due to the fact that embedding ESDs will
likely be extremely inefficient, embedded event selection for ESDs was
disabled. Suggestions for how to re-enable it would be appreciated!