EMCal Embed: Update jet base classes [ALPHY-77]
Updates the classes to avoid passing the wrong TClonesArray when using
multiple EMCal containers and accessing track and cluster constituents.
Note that these changes basically remove the dependence of
AliJetContainer on particle and cluster containers. They could likely
be removed in the future.

For more information, see the JIRA ticket ALPHY-77