Disable Weffc for gcc for AliV0ReaderV1
gcc implements the class std::iterator to have a non-virtual
destructor. Consequently classes inheriting from std::iterator
will raise the Weffc warning. Inheriting from std::iterator
however is a popular and easy way to make a class std::iterable,
and the way chosen in this approach. As the std::iterator doesn't
allocate memory, the warning is still vaid but harmless, and
as it cannot be handled by any implementation it should be treated
as a false positive. Therefore Weffc will be disabled for gcc
for AliVOReaderV1 in a similar way is it is done for the iterator
of ROOT's TMap.