[ALPHY-81] Fix primary vertex position handling in cluster container
The position of the primary vertex has to be set per event. Instead it was
handled in SetArray, which is only called in ExecOnce. A setter function for
the primary vertex was added, and it is called automatically in AliAnalysisTaskEmcal
in function run. Affects momentum vector calculation for neutral constituents in
AliEmcalJetTask (jetfinder task of the EMCAL jet framework).
(see https://alice.its.cern.ch/jira/browse/ALPHY-81 for more details)