Introduce YAML config class
Introduces a new configuration class, AliYAMLConfiguration,
that provides an interface which allows users to read and
write values in YAML files. It is refactored and generalized
from the YAML functionality present in the EMCal Corrections
Framework.

Given the generalization, it was moved to PWG/Tools, with the
hope that it can benefit other users. Note that files stored
on AliEn are fully supported.

Documentation on usage is available in the class header.