Fixing operator ambiguity
Due to the implementation of the casting operators and
arithmetic operators with int/float types in TComplex,
there is an ambiguity on whether to use
- operator- from TComplex with TComplex and int
- operator- from int with TComplex implicitly casted to int
In order to fix the ambiguity the complex value needs to
be explicitly defined, or the integer can be converted to
complex using the value as real part of the complex number.
The proposed solution does the subtraction on complex level.

Needed for ROOT6 builds.