Implementation of a constituent wrapper class
Goal of the development is
- Provide direct access to constituent objects, without passing
  arrays to it (a reconstructed object has to know about its childs)
- Adding meta information to the constituent object: This is in
  particular of relevance for the clusters and the energy definition
  used in the jetfinder.
The wrapper class is implemented to be minimal, keeping only the pointer
and (if needed) additional minimal data entries. Constituents are stored
in prarallel to the old indices in order to keep backward compatibility,
in a TClonesArray. As result the containers for constituent objects become
std-iterable.