Test code for new constituent based handling (experimental)
Constituent based handling (constituent class serving as smart pointers
to underlying objects) to be committed soon. In preparation for this
a small test suite checking that the old (index-based) and new (pointer-based)
implementation are the same. The test is protected with the preprocessor define
EXPERIMENTAL_JETCONSTITUENTS, so it will only be compiled in the experimental
version of the code.