EMCal CorrFW: Add additional config on LEGO train
New function to retrieve an existing correction task so that it can be
configured further in a different wagon on the LEGO train.

To achieve this, a dummy task is created when the configure task is
called because AliAnalysisAlien requires that all wagons add a task.
Then, when Initialize(true) is called on the correction task, the dummy
task is removed.