Silence pthard mismatch handling for old samples
pt-hard bin not encoded in path name of old productions,
therefore check for the correct pt-hard bin fails. This
check can now be switched off by the user, in this case
the pt-hard bin is not determined from the file name and
the set to bin 0 automatically. In addition a flag that the
pt-hard bin is not initialized is set, for which the check
is disabled.

Users can switch off the automatic pt-hard bin detection via

AliAnalysisTaskEmcal::SetGetPtHardBinFromPath(false);

In case the automatic pt-hard bin detection fails (i.e. disabled),
also the histograms after selection are filed with bin 0 in order
to aviod mismatch of the pt-hard bin before and after selection,
even though the pt-hard bin can be obtained correctly from the
pt-hard value stored in the event.