JetH: Support measuring embedded data
Adds support for measuring external event hadrons correlated with
a jet collection. In the embedded configuration defined in the
class, it correlates the hadrons from the external event with jets found
in the combined data of the external event embedded into the internal
event.

It also includes improved debug information