EMCal CorrFW: Track to container map for had corr
Creates a mapping for tracks to their corresponding container in the
hadronic correction when running with AODs. AOD clusters only
provide pointers to tracks, so this map is necessary to select the
proper container, which allows the proper track
selection to be applied to each track. This is important if the
containers have different track selection conditions (as might happen
in, for example, embedding pp data into PbPb).