[EMCAL-42] Implementation of equalty and streaming operators
Equalty operator:
Use equalty operators in order to determine whether two DCS configuration
objects are the same. Objects are defined as equal if
- For STU/TRU object all parameters are the same
- For the DCS config both objects have the same TRUs and STUs, and the
  corresponding configs are the same
For STU configs the error counters are not considered in the check for
equalty

Streaming operators:
Implemented for TRU and STU config, putting all the parameters on the
stream. In case of STU, error counters - if available - are not put on
the stream.