Abstractize underlying handling of the track selection
In order to merge AliEmcalTrackSelectionESD and
AliEmcalTrackSelectionAOD the track selection method
must rely on using the same interface function (currently
AliVCuts::IsSelected). Therefor various special treatment
for both cases AOD / ESD has to be expressed as an AliVCuts
object properly implementing the IsSelection method. The
following modifications are relevant for this:
- AliEmcalAODHybridTrackCuts: Implementation of the selection
  of hybrid tracks via AliAODTrack::IsHybridGlobalContrainedGlobal
- AliEmcalESDHybridTrackCuts: Implementation of the hybrid track cuts
  for the case  of AliESDtracks (various options)
- AliEmcalESDtrackCutsWrapper: Work-around for the improper
  implementation of the function IsSelected of AliESDtrackCuts
  in case of AOD tracks
As ususal for new classes the new classes are put into the
namespace PWG::EMCAL.