Add possibility for user-defined return values in the track selection
Allow passing user informtion from the track seleciton (i.e. hybrid track
status) to user consumers (i.e. track selection). For this a new cut
interface is needed, which defines the function IsSelected in a way that
it returns a smart pointer containing the selection status and optionally
a user-defined object (cut class: AliEmcalCutBase). Memory management of
the user-defined object is done using reference counting. Handling of cut
objects deriving from AliVCuts is made transparent with the help of a wrapper
(AliEmcalVCutsWrapper) - in this case the result pointer has no user data
associated.
The track selection can handle multiple cuts at the same time. For this
the return type must be an object providing access to the selection results
of all track selections (AliEmcalTrackSelResultCombined). A first
implementation of a user-defined object is provided for hybrid tracks
(AliEmcalTrackSelResultHybrid).