Express TPC-only track selection as virtual cut class
Implementation of a virtual cut class AliEmcalAODTPCOnluTrackClass
for the selection of TPC-only hybrid tracks. All selections on
AOD level are now expressed as cut objects in the same way as
done on ESD level, and the IsTrackAccepted functions are now the
same and can be combined at a later stage.