PWGGA/GammaConv: Implementation of Correction Framework to Calo Tasks
This commit implements the correction framework from the EMCal group
to the PWGGA/GammaConv calorimeter based tasks (Calo, ConvCalo and
CaloMerged) as well as the necessary analysis cut classes. The AddTasks
for pp have already been modified for this change. For pPb and PbPb the
changes can be made if desired.