Fix needed to overcome double-delete issue with ROOT v6-12-04
It turned out that the double-delete issue with ROOT v6-12-04
appears for objects, in particular inheriting from TCollection,
which are bound to a TFile but not delete in the task destructor.
Once deleting them in the destructor the crash disappears.

This affects the destruction at the end of the ROOT session and
works as long as there is only one task writing to the same output
file. As soon as there is a second instance of a task (same task
or different) writing to the same output file the segfault reappears.