EMCal Embed: Fix combining cell collections
There were some indexing bugs which prevented the cell collections from
being combined successfully. Those were fixed, with new functionality
added to verify they were successfully combined.

Note that the length of the combined array should always been the same
length as the combination of the lengths of the cell objects, but that
doesn't mean that the arrays are the same. This can only be verified by
checking each cell by hand.