Fix doxygen typos and issues
The first `//` is eaten by doxygen in the EMCal corrections readme for
no apparent reason (as the exact same type of code is shown directly
below without the `//` being lost), so they are doubled up to ensure
that the comment is displayed properly when the doxygen is compiled.

Also fixes a few typos.