Modular jet substructure tree
Flat array no longer used. Instead the data is organized
in structures (partially using the structures used as
return data for the SoftDrop/N-subjettiness). Each
structure has a function LinJetTreeBranch to link its
variables to a branch in the output tree. The structures
are dynamic data members of the class, allocated in
UserCreateOutputObjects and deleted in the constructor.