EMCal Embed: Add internal event selection
Previously, good embedded events were lost when the internal event
was outside of, for example, the desired centrality range. Now,
the user should check for
AliAnalysisTaskEmcalEmbeddingHelper::GetIstance()->EmbeddedEventUsed()
to ensure that the internal event passes event selection and that
the embedded event was embedded successfully. If the internal event
did not pass event selection, the embedded event is saved for the
next internal event.