EMCal Embed: Move file access check to Initialize
When file access was checked in SetupInputFiles(), it would cause
unnecessary load to the grid due to AccessPathName calling a command in
xalienfs (`gapi_stat()`), which then performed an `ls -l` on every
single directory. If we instead move it to Initialize(), such checks
will only have to be performed once.