Add track efficiency options, physics sel fixes
- Allow setting of track eff regardless of beam type
- Add artificial tracking inefficiency option. Operates the same way as
in AliEmcalJetTask, with the exception that a list of rejected indices
is maintained to ensure that the random rejection process is only
applied once, such that those randomly rejected particles are always
for a particular event.
- Simplify handling of track inefficacy, ensuring a list will be used
after the tracks have been looped over once
- Properly handle physics selection of 0. Note that 0 & 0 == 0, and thus
MC physics selection (where 0 is a valid selection) needs to be handled
specifically. It can be handle by checking the trigger mask with an or
for if the values are identically equal (ie if both are 0).