EMCal Embed: Fix embedded physics selection
Embedded physics selection was not considered if the default value of kAny
was set. However, this causes problems if the physics selection is set to
0 (a valid selection for MC), in which case the physics selection is
applied, but no events are ever selected because `0 & 0 == 0`.

To resolve this issue, the default physics selection is now 0. In the case
that 0 is still the value when attempting to apply physics selection, it
will not be applied, thereby ensuring that events with 0 can
successfully be embedded. Practically, this should not change any
results because kAny was almost always used, implicitly ensuring that no
physics selection was applied. However, with this fix, it is now
important to have the proper physics selection!