EMCal Embed: Update jet taggers
The tasks use `AliEmcalJet::TrackAt(int index, TClonesArray * tracks)`,
with the particles accessed via the particle container attached to the
jet container. This will cause harmless warnings (the proper particle
is _always_ returned) as explained in the docs of `AliEmcalJet` and
of the embedding framework. However, the tasks should be updated to
silence these warnings under the new embedding framework.

This update causes no change in functionality.

Two compiler warnings for unused variables are also fixed.