Herwig: Automatic matchbox file generation from template with JINJA
Matchbox files contain several keys which need to be handled on-
the-fly (number of events, random seed). The easiest solution
to configure matchbox file from static and dynamic settings is
treating Matchbox.in files as JINJA templates with the JINJA
conventions and use a small python script called in gen_herwig.C
to create the Matchbox.in file from the Matchbox.tmpl file.
For configuration settings the python script assumes the AliDPG
naming convention (global variables, starting with CONFIG_).

Attention: The script should be called on the workers. Requires
the python package JINJA to be installed on the grid sides. The
commit here represents only an example at the current stage and
might not work on every grid site.