[EMCAL-135] Disable temperature recalib due to memory leak
It has been observed that the tender and the correction framework
leak ~100 MB once a run is changed, leading to problems for jobs
processing multiple runs. 70% of the leaking is produced from
loading of the temperature recalibration. As the temperature
recalibration is not used in run2 initializing the recalibration
factors is disabled in order to prevent opening the OADB file untill
the memory leak in the OADB container is fixed.