[EMCAL-135] Use std::unique_ptr to handle lifetime of OADB containers
Addressing memory leak discussed in EMCAL-135: InitFromFile
of AliOADBContainer dropped in favour of automatic lifetime
management of the OADB containers using std::unique_ptr.
Depends on the fix in AliOADBContainer setting the ownership
of the object and the fix in AliEMALRecoUtils taking the
ownership of calibration data.

In addition: Renaming of local variables as requested by
@raymondehlers in #4792