[EMCAL-172, ALIROOT-7718, ALIROOT-7649] Disable trigger simulation via reco params
Trigger electronics only createed in case it is enabled
in the EMCAL reco params (new field in reco params with
default value true). In case the trigger electronics is a
nullptr (not constructed in the constructor) the trigger
simulation will not be processed.

Lazy evaluation is used for the initializaiton of the trigger
electronics: Initialization is put into the function
InitTriggerElectronics which is called in FillESDs. In case
the trigger electronics is initialized the function does
nothing. Otherwise it will try to initialize the function
in case it is enabled in the reco params. Since a new state
for disabled trigger electronics is introduced with this commit
a check for nullptr is not possible to identify initalization
status because nullptr is an allowed status. Therefore the
function will check whether to initialize the trigger electronics
every time the function is called. This method is necessary
because the recoparams are not available neither at constructor
nor at init level.

Off-status for STU is handled via the DCS region: For OFF
STUs the DCS sends  null values for all entries of the
STU configuration. The off-status can therefor be deduced
from a 0 value in the STU region. In this case the active
status of the STU is set to 0, and for inactive STUs the
function L1 processing the L1 trigger simulation will be
skipped.