Reverse order of retrieving values from config
If values are retrieved in the order that the configuration files are
added, then it is impossible to override a value after a file has been
added. Instead, we now look in reverse order of the way they were added,
so a value can be overridden just be adding a new config file.