Add support for virtual trigger selection
Tasks inheriting from AliAnalysisTaskEmcal can
reimplement the function IsTriggerSelected to have a
custom trigger selection applied during the event
selection. The default implementation of the function
is identical to the previous trigger selection part in the
IsEventSelected function in order to guarantee backward
compatibility. Custom trigger selection is in particular
necessary for the EMCAL triggers when calculating the
vertex efficiency.