[EMCAL-369] Cache ALTRO mapping in AliHLTCaloRawAnalyzerComponent
The initialization of the ALTRO mapping accounts for a significant amount
of the time spent in the raw analyzer component, as the mapping is read
for every event and DDL from an ASCII file. Instead the mapping can be
cached in the component and passed to the AliCaloRawStreamV3 as
external mapping. The code extracting and caching the mapping is
copy&pasted from AliCaloRawStreamV3. The mapping is initialized the first
time a calo raw stream is created.