Ignore FPE when calibrating TRD
The error appears on ROOT 6 and makes the TRD calibration to stop under certain
conditions. SIGFPE has different types, the one typically raised here is an
"inexact result", when the rounded result returned by the processor is different
from the infinitely precise result.

It should be safe to ignore such exception in this context.