EMCal CorrFW: Default same crosstalk and min cellE
By default, we want the minimum induced energy in a T-card to be the
same as the minimum cell energy in the clusterizer. To modify one value
on it's own, change the individual value directly to the parameter (or
to change both, change the shared parameter).

Explicitly, to change the T-card minimum to 0.1 while leaving the
clusterizer minimum cell energy the same in your user config, it should
look like (omitting unrelated values for brevity):

```yaml
sharedParameter:
    minCellE: 0.05
EmulateCrosstalk:
    inducedTCardMinimum: 0.1
Clusterizer:
    minE: "sharedParameters:minCellE"
```

To change the minCellE for both in your user config, you'll need to do:

```yaml
sharedParameter:
    minCellE: 0.05
EmulateCrosstalk:
    inducedTCardMinimum: "sharedParameter:minCellE"
Clusterizer:
    minE: "sharedParameters:minCellE"
```

It isn't enough to just change the value in the shared parameter!

For more information, see the documentation in
`PWG/EMCAL/READMEemcCorrections.txt`.